troubleshoot react-router update to v6 way of routing
