+++ src/App.tsx
@@ -2,20 +2,26 @@ import {UkraineButton} from '@gtibrett/mui-additions';
 import {Box, CssBaseline, useTheme} from '@mui/material';
 import React from 'react';
 import {Provider} from 'react-redux';
-import {BrowserRouter} from 'react-router-dom';
+import {BrowserRouter, Outlet, Route, Routes, useLocation} from 'react-router-dom';
+import ErrorBoundary from './app/ErrorBoundary';
 import Header from './app/Header';
-import Routes from './app/Routes';
+import useRoutes from './app/useRoutes';
 import {ThemeProvider} from './components';
 import store from './redux/store';
 
 const InnerApp = () => {
-	const theme = useTheme();
+	const theme    = useTheme();
+	const location = useLocation();
+	
+	console.info({location});
 	
 	return (
 		<Box sx={{position: 'fixed', top: 0, left: 0, bottom: 0, right: 0, background: theme.palette.background.default, overflow: 'auto'}}>
 			<Header/>
 			<main>
-				<Routes/>
+				<ErrorBoundary>
+					<Outlet/>
+				</ErrorBoundary>
 				<UkraineButton/>
 			</main>
 			<footer/>
@@ -24,12 +30,18 @@ const InnerApp = () => {
 };
 
 export default function App() {
+	const navLinks = useRoutes();
+	
 	return (
 		<Provider store={store}>
 			<CssBaseline/>
 			<ThemeProvider>
 				<BrowserRouter>
-					<InnerApp/>
+					<Routes>
+						<Route path="/" element={<InnerApp/>}>
+							{navLinks.map(({label, ...route}) => <Route {...route}/>)}
+						</Route>
+					</Routes>
 				</BrowserRouter>
 			</ThemeProvider>
 		</Provider>
